+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 The structure and content of this file follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.14.3] - 2022-06-12
+### Fixed
+- Returned `[]byte` from oj.Marshal and pretty.Marshal now copy the
+  internal buffer instead of just returing it.
+
 ## [1.14.2] - 2022-06-03
 ### Added
 - Added SameType test tool.